Improve conditional inclusion of tracing, also condition on Proton version
