+++ src/api/qpid-proton/common/ModernClient.h
@@ -13,22 +13,6 @@
 #include <string>
 #include <memory>
 
-#include <proton/tracing.hpp>
-
-#ifdef __unix__
-#include <opentelemetry/sdk/trace/simple_processor.h>
-#include <opentelemetry/sdk/trace/tracer_provider.h>
-#include <opentelemetry/trace/provider.h>
-#include <opentelemetry/nostd/unique_ptr.h>
-#include <opentelemetry/exporters/jaeger/jaeger_exporter.h>
-#include <opentelemetry/exporters/ostream/span_exporter.h>
-#include <opentelemetry/sdk/resource/resource.h>
-
-#include <opentelemetry/trace/span.h>
-#include <opentelemetry/trace/tracer.h>
-#include <opentelemetry/trace/context.h>
-#endif
-
 #include "Client.h"
 #include "logger/Logger.h"
 #include "logger/LoggerWrapper.h"
