+++ src/api/qpid-proton/common/ModernClient_tracing_none.cpp
@@ -1,0 +1,8 @@
+#include "ModernClient.h"
+
+void ModernClient::enableTracing(std::string service_name) const
+{
+
+    // do nothing
+
+}
